src/test/java/org/hanghae/markethub/domain/purchase/service/PurchaseServiceTest.java
package org.hanghae.markethub.domain.purchase.service;

import org.hanghae.markethub.domain.cart.entity.Cart;
import org.hanghae.markethub.domain.cart.repository.CartRepository;
import org.hanghae.markethub.domain.item.entity.Item;
import org.hanghae.markethub.domain.item.repository.ItemRepository;
import org.hanghae.markethub.domain.purchase.dto.PurchaseRequestDto;
import org.hanghae.markethub.domain.purchase.dto.PurchaseResponseDto;
import org.hanghae.markethub.domain.purchase.entity.Purchase;
import org.hanghae.markethub.domain.purchase.repository.PurchaseRepository;
import org.hanghae.markethub.global.constant.Status;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;

import java.util.ArrayList;
import java.util.List;
import java.util.Optional;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.api.Assertions.*;
import static org.mockito.ArgumentMatchers.any;
import static org.mockito.Mockito.verify;
import static org.mockito.Mockito.when;

class PurchaseServiceTest {

    @Mock
    private CartRepository cartRepository;

    @Mock
    private PurchaseRepository purchaseRepository;

    @InjectMocks
    private PurchaseService purchaseService;

    @Mock
    private ItemRepository itemRepository;

    @BeforeEach
    void setUp() {
        MockitoAnnotations.openMocks(this);
    }

    @Test
    @DisplayName("주문 생성 로직 테스트")
    void 주문하기테스트() {
        // Given
        String userId = "testUser";
        Status status = Status.EXIST;
        PurchaseRequestDto requestDto = new PurchaseRequestDto(status);

        List<Cart> carts = new ArrayList<>();

        when(cartRepository.findByUserId(userId)).thenReturn(carts);
        when(purchaseRepository.save(any(Purchase.class))).thenAnswer(i -> i.getArgument(0));

        // When
        PurchaseResponseDto responseDto = purchaseService.createOrder(requestDto, userId);

        // Then
        assertNotNull(responseDto);
        assertEquals(responseDto.status(), status);
        assertEquals(responseDto.carts().size(), carts.size());
        verify(cartRepository).findByUserId(userId);
        verify(purchaseRepository).save(any(Purchase.class));
    }

    @Test
    @DisplayName("단일 아이템 주문 등록 테스트")
    void 단일아이템주문테스트(){
        Long itemId = 1L;
        Status status = Status.EXIST;
        PurchaseRequestDto.SinglePurchaseRequestDto requestDto = new PurchaseRequestDto.SinglePurchaseRequestDto(status, itemId);
        Item item = new Item(); // 적절한 아이템 객체 생성

        when(itemRepository.findById(itemId)).thenReturn(Optional.of(item));
        when(purchaseRepository.save(any(Purchase.class))).thenAnswer(i -> i.getArgument(0));

        // When
        PurchaseResponseDto responseDto = purchaseService.createSingleOrder(requestDto);

        // Then
        assertNotNull(responseDto);
        assertEquals(responseDto.status(), status);
    }

    @Test
    @DisplayName("단일 아이템 주문 예외처리 테스트")
    void 단일아이템주문예외처리테스트() {
        Long itemId = 1L;
        Status status = Status.EXIST;
        PurchaseRequestDto.SinglePurchaseRequestDto requestDto = new PurchaseRequestDto.SinglePurchaseRequestDto(status, itemId);

        when(itemRepository.findById(itemId)).thenReturn(Optional.empty());

        // When & Then
        assertThrows(IllegalArgumentException.class, () -> {
            purchaseService.createSingleOrder(requestDto);
        });
    }

    @Test
    @DisplayName("이메일로 모든 구매 내역 조회 테스트")
    void findAllPurchaseByEmailTest() {
        // Given
        String email = "user@example.com";
        List<Purchase> mockPurchases = List.of(new Purchase(Status.EXIST, null)); // 테스트 데이터 준비
        when(purchaseRepository.findAllByUserEmail(email)).thenReturn(mockPurchases);

        // When
        List<PurchaseResponseDto> responseDtoList = purchaseService.findAllPurchaseByEmail(email);

        // Then
        assertNotNull(responseDtoList);
        assertFalse(responseDtoList.isEmpty());
        assertEquals(mockPurchases.size(), responseDtoList.size());
        verify(purchaseRepository).findAllByUserEmail(email);
    }

    @Test
    @DisplayName("이메일로 단일 구매 내역 조회 테스트")
    void findPurchaseByEmailTest() {
        // Given
        String email = "user@example.com";
        Purchase mockPurchase = new Purchase(Status.EXIST, null); // 테스트 데이터 준비
        when(purchaseRepository.findByUserEmail(email)).thenReturn(mockPurchase);

        // When
        PurchaseResponseDto responseDto = purchaseService.findPurchaseByEmail(email);

        // Then
        assertNotNull(responseDto);
        assertEquals(mockPurchase.getStatus(), responseDto.status());
        verify(purchaseRepository).findByUserEmail(email);
    }
}